+++ packages/config/vitest.config.ts
@@ -1,5 +1,5 @@
-import { defineConfig } from "vitest/config";
+import { defineProject } from "vitest/config";
 
-export default defineConfig({
+export default defineProject({
   test: { pool: "forks" },
 });